requirements.txt
Flask
Flask-SQLAlchemy
requests
kafka-python